Working on Roles pointer error
