+++ CodeTests/Code Results/RolesResults.txt
@@ -6,17 +6,19 @@ Stab
 - Type: Physical Attack
 - Effect: 52
 
+Starting
 Printing Stab to file...Done
 
+Starting
 Printing FireBlast to file...Done
 
 Stab
 - Type: Physical Attack
 - Effect: 52
 
-0xdd1a10
-0xdd1a10
-0xdd1a10
+0x801510
+0x801510
+0x801510
 Stab has been added to Mage role.
 
 Adding a New Power to the Mage Role.
@@ -101,14 +103,18 @@ Light Armor 0
 Powers List
 ---------------
 Power 1: Stab
+Starting
 Printing Ice Blast to file...Done
 
+Starting
 Printing Chaos to file...Done
 
 Printed Role: Mage to file.
+Starting
 Printing Strike to file...Done
 
 Printed Role: Warrior to file.
+Starting
 Printing Stab to file...Done
 
 1
