+++ TestResults.txt
@@ -1,13 +1,12 @@
 Gold Cap
 Gold Chestplate
 Gold Sword
-Gold Knief
 
 
 
 Gold Cap
 ------------------------ 
- Type: Medium Helmet
+ Type: Broken ItemMedium Helmet
  Level: 5
  Description: Simple persons hat
  Ability Points: 0 DEF
@@ -15,7 +14,7 @@ Gold Cap
 
 Gold Chestplate
 ------------------------ 
- Type: Medium Chestplate
+ Type: Broken ItemMedium Chestplate
  Level: 5
  Description: Simple persons chestplate
  Ability Points: 0 DEF
@@ -23,17 +22,9 @@ Gold Chestplate
 
 Gold Sword
 ------------------------ 
- Type: Sword
+ Type: Broken ItemSword
  Level: 5
  Description: Simple persons sword
- Ability Points: 0 ATK
-
-
-Gold Knief
------------------------- 
- Type: Knief
- Level: 6
- Description: Simple persons knief
  Ability Points: 0 ATK
 
 
@@ -48,17 +39,15 @@ Equipped: Gold Chestplate
 
 Equipped: Gold Sword
 
-Equipped: Gold Knief
-
-2
+1
 ARMOR
 -----------------------
 Head: Gold Cap
 Chest: Gold Chestplate
-Weapon: Gold Knief
+Weapon: Gold Sword
 Pants: This Slot is Empty
 
-Unequipped Gold Knief.
+Unequipped Gold Sword.
 
 ARMOR
 -----------------------
@@ -83,7 +72,7 @@ ARMOR STATISTIS
 Head
 Gold Cap
 ------------------------ 
- Type: Medium Helmet
+ Type: Broken ItemMedium Helmet
  Level: 5
  Description: Simple persons hat
  Ability Points: 0 DEF
@@ -91,7 +80,7 @@ Gold Cap
 Chest
 Gold Chestplate
 ------------------------ 
- Type: Medium Chestplate
+ Type: Broken ItemMedium Chestplate
  Level: 5
  Description: Simple persons chestplate
  Ability Points: 0 DEF
@@ -99,7 +88,7 @@ Gold Chestplate
 Weapon
 Gold Sword
 ------------------------ 
- Type: Sword
+ Type: Broken ItemSword
  Level: 5
  Description: Simple persons sword
  Ability Points: 0 ATK
@@ -130,3 +119,17 @@ Opening File: ./Data/TempEList.txt
 Printed Gold Cap
 Printed Gold Chestplate
 Printed Gold Sword
+Done
+Printed to File
+
+===============================
+Role: Mage
+Armor Type: Light Armor
+-----Role Powers-----
+	There are no powers.
+
+
+ 
+There are no powers.
+Deleting Role: Mage
+Deletion Done
